requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
FLAML==2.2.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
openai==0.28.1
psycopg2-binary==2.9.8
pyautogen==0.1.7
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
